feat: split clashcore (1.0.0) out of clashmap (1.3.0)
Move the lock primitives, detached guards, sharded collection, and
Ref/RefMut/RefMulti/RefMutMulti types into a new `clashcore` crate
(version 1.0.0) and re-export them from `clashmap` (now 1.3.0) so the
1.x public API is unchanged.

`clashcore` carries the lock_api / parking_lot_core / polonius-the-crab
dependencies; `clashmap` only keeps hashbrown, crossbeam-utils, and
replace_with. The previously private `pub(crate)` items that now cross
the crate boundary (Ref/RefMut::new + into_parts, ClashCollection::shift
/ from_parts / try_fold, RwLock*GuardDetached, try_map) are promoted to
`pub`; their `unsafe` markers and safety contracts are unchanged.

The `#[no_mangle]` previously placed on `default_shard_amount` is
dropped — it was unintentional and clashes with the goal of having two
separate crates expose related items.

Tests that asserted on `ClashTable` move with `clashmap` rather than
`clashcore`; lock-level tests stay alongside the lock implementation in
`clashcore`. `cargo build/test/doc/clippy --workspace --all-features`
all pass.
