+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["crates/clashmap"]
+members = ["crates/clashcore", "crates/clashmap"]
 resolver = "2"
